backend/requirements.txt
fastapi[standard]
uvicorn
python-multipart
pypdf
python-docx
python-dotenv
openai
pandas
openpyxl
matplotlib
